backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
python-dotenv==1.0.1
firebase-admin==6.5.0
cloudinary==1.40.0
python-multipart==0.0.9
openai==1.30.5
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
python-slugify==8.0.4
pypdf==4.2.0
